Full Yeeeee

+++ src/main/java/pl/akalkowski/demo/models/Category.java
@@ -15,7 +15,7 @@ public class Category {
 
 
     @GeneratedValue(strategy = GenerationType.AUTO)
-    Integer id;
+    Long id;
     String name;
 
     public Category() {
@@ -27,16 +27,16 @@ public class Category {
         name=form.getName();
     }
 
-    public Category(Integer id, String name) {
+    public Category(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
